+++ test/argus/dl_declarations_test.exs
@@ -137,11 +137,11 @@ defmodule Argus.DlDeclarationsTest do
       deferred_startup_deadlock:
         ~w(async_cast call_arg call_arg_forward call_edge call_tag callback_tag callback_timeout dynamic_child function_def handle_continue_clause implements_behaviour init_continues_to supervisor supervisor_child sync_call try_start),
       distributed:
-        ~w(call_edge catch_falls_through catch_handler catch_tag distributed_store_op function_def global_op global_register implements_behaviour node_operation rpc_call try_call),
+        ~w(call_edge catch_falls_through catch_tag distributed_store_op function_def global_op global_register implements_behaviour node_operation rpc_call try_call),
       error_handling:
-        ~w(bare_rescue call_edge callback_tag callback_total catch_handler catch_tag catch_total exit_call function_def ignored_error_result implements_behaviour monitor_call trap_exit try_call),
+        ~w(bare_rescue call_edge callback_tag callback_total catch_tag catch_total exit_call function_def ignored_error_result implements_behaviour monitor_call trap_exit try_call),
       ets:
-        ~w(call_edge catch_handler closure_def dynamic_child ets_new ets_op ets_option function_def implements_behaviour statem_event_clause supervisor_child),
+        ~w(call_edge catch_tag catch_total closure_def dynamic_child ets_new ets_op ets_option function_def implements_behaviour statem_event_clause supervisor_child),
       gen_statem:
         ~w(function_def statem_call_unreplied statem_event_catchall statem_event_clause statem_info_catchall statem_initial statem_module statem_state statem_timeout statem_transition),
       message_contract: